My orders details page UI with dynamic container is done

+++ lib/models/MyOrdersDataModel.dart
@@ -9,7 +9,7 @@ class MyOrdersDataModel {
   int? quantity;
   List<CartItemsModel>? orderedItems;
   double? subTotal;
-  int? shippingCharges;
+  double? shippingCharges;
   double? total;
   String? deliveryStatus;
